requirements.txt
certifi==2020.12.5
chardet==3.0.4
idna==2.10
pydub==0.24.1
python-dotenv==0.15.0
requests==2.25.0
urllib3==1.26.2
